requirements.txt
Django>=4.2
python-dotenv
django-admin-interface
django-flat-theme
requests
requests
